Add GeographicSearch concern
GeographicSearch provides scope '.close_to' which takes arguments for
coordinates and a search radius.

PostGIS (and all GIS utilities) work in metres by default so the
conversion from miles is made. This will be moved to its own module in
an upcoming commit but as everything currently works it's included here.