Switched to using PostGIS enabled version of Postgres
Within the Azure Pipeline, we run Postgres from a Docker Image

The official image does include support the PostGIS extension within the Docker
image, so `create extension postgis` will always fail.

Switching to mdillon's fork which includes the PostGIS extension. This has 10M
downloads at this point and seems to be regularly updated. The running container
is not accessible outside of the internal Docker network within the CI
environment so this seems safe to depend upon from a security perspective
despite being a non official image.