Restructured Bookings::SchoolSearch, added sorting
Now the service object is initialized with the search arguments and the
results method actually executes the query. This was changed to cater
for the additional step that only allows distance to be set as a sort
option if we know where the candidate is.

The search results also include a distance attribute (in metres!) that
can be displayed with the results.