Support multipart date params
Multipart date params don't seem to be supported by ActiveModel.
We include `ActiveModel::Attributes` to get this behaviour, however
doing so requires us to patch ActiveModel::Type::Date to define a
default timezone. For now we're defaulting to UTC, but the timezone
choice shouldn't matter as we're only submitting dates.