Adds base class for registration wizard steps
Adds `persisted?` which we'll want for setting the method on form_fors.
Also adds created_at and updated_at time stamps. We need something to
determine if a model is persisted so may aswell use a timestamp. Also
can compare updated_at and created_at to capture information on if a
particular step is often edited.