Add helper to make elements visible from Selenium
The GOV.UK Design System replaces checkboxes[0] and radio buttons[1]
with a thick-bordered element that acts like the actual input, which is
set to `opacity: 0`. Selenium, which mimics an actual user cannot
interact with 'invsible' items so fails.

This change runs a snippet of JavaScript that, whenever a page is loaded
using a Selenium-based driver, resets the opacity of the input to 1
making it fully-opaque

[0] https://design-system.service.gov.uk/components/checkboxes/
[1] https://design-system.service.gov.uk/components/radios/