Use Redis directly.
Changes registration session store to use Redis directly rather than
going through Rails.cache.
We're using Rails.cache to store the session data in Redis but this is
an implementation detail we may want to change in future.
We want to differentiate between the session data and the PII we're
temporarily storing in Redis before writing to postgres (once the user
has confirmed their email address).