Add rake task assets:symlink_non_digested
This rake task, which is a modified version of the one found in the
comments of a Sprockets bug[0], creates a symlink of all digested CSS
files without the hash. This allows us to reference them from static
error pages.

So, in `public/assets`, a symlink like the following would be created:

application.css -> application-16b44...5122.css`

[0] https://github.com/rails/sprockets-rails/issues/49#issuecomment-20535134