makes PlacementPreference#availability required
Adds availability as a required field for PlacementPreferences and
updates copy to bring the app in line with the latest changes to the
prototype.
Pulls in changes from updated Notify template.
Bumps form builder version as we want the multi hint support and the
ability to overwrite the default class on labels.