Reinitialise state in GrabLocationController js
When using the 'use my location' feature, then performing a search, followed by
navigating back to the original page, the widget ends up in an inconsistent
state.

The text fields are repopulated, the hidden fields are not.

Make the widget reinitialise itself properly on load and detect the inconsistent
state between text and hidden fields and just reset to back to unset.