Scope registration sessions by school urn
Previously if a user performed two simultaions registrations for
different schools in seperate windows the first registration session
would be overwritten by the second. This commit fixes that issue by
scoping the session by school urn.
School session is not needed for registration store as the store
generates a uuid for each registration session.