Moves magic link uuid from params to session.
There was an issue with redirecting to confirmation_emails#show with the
uuid in the url such that the user could skip straight to
placement_requests/new?uuid=<uuid> without having to click the link in
the email. This PR fixes the issue by storing the uuid in the session
when redirecting to confirmation_emails#show.

The instances of RegistrationSession are now responsible for generating
a uuid rather than the RegistrationStore. The store now asks the session
for it's uuid.