Fix missing error border on textarea with errors
The form builder adds .govuk-input--error rather than
.govuk-textarea--error to textareas with error. However the javascript
that provides the maxwords behaviour removes .govuk-textarea--error from
textareas on page load (it adds this class when the word count is too
high and removes it once the word count is low enough). The textarea's
normal border overrules .govuk-input--error border (stylesheet is
arranged alphabetically) so we need to add this styling with heigher
specificity.