Populate new form with session attributes
If canidates come back to the registration wizard after completing some
steps we'll prepopulate the new form with the information they've
entered so far so they can more quickly get back to the last step they
were on.

In `#attributes_from_session` we need to remove created_at to avoid the
form submitting to the update action.