Modify the compare task to show the title too
This is now required because moving the templates to the official Notify
account means emails that aren't used by the application are included in
the results. There doesn't appear to be a way of making the response
tell us which folder/directory templates are in, so there's no reliable
way of omitting unrelated templates.

This change adds the template's name to the GUID and status in the
output in an attempt to make the reconciliation process simpler