Adds school on boarding candidate_requirements
Adds the candidate requirements step of the school on boarding wizard
and the on boarding wizard setup.
The wizard will build up a school_profile with each screen backed by a
form model and the school_profile being composed of these form models.
This allows for a seperate model for each end point so we can play as
nicely with Rails as possible.