Lock to Ruby 2.5.3
Gemfile is currently locked to a specific Ruby version. Either we slacken that
requirement or increase strictness of Dockerfile.

Since 2.5.4 has been released with known problems with Puma, and since 2.5.5
which fixes the issue, is not yet available in the Docker registry - I'm locking
to a specific version and taking the view that changing versions is something
which should be a code change by the developers.