Populate subject choice with schools subjects if it has any
If the school has provided us with a list of subjects during enrichment
we'll show that list, if the school has not we'll default to showing all
available subjects.

Testing:
Find a school which has it's own subjects, eg urn 142541,
complete the wizard up to the "We need some more details" screen
expect the list of subjects in the what subject do you want to teach selects to
match the schools' subject.

Find a school which has no subjects, eg urn 138304,
complete the wizard up to the "We need some more details" screen
expect the list of subjects in the what subject do you want to teach selects to
show all subjects.