Disabled the find-my-location field on Internet Explorer
Whilst Internet Explorer implements the navigator.geolocation feature. It
doesn't implement it fully, missing the 'failed to get location' callback.
This means if the user disables the prompt, then we never receive a message to
say the browser has stopped searching for the location.

Since the feature is an optional enhancement, IE only partially implements this
feature, and the use case for supporting it under IE is very niche, I am
instead just disabling the feature under Internet Explorer.