Add expanded search area to results
If no results are found the radius is expanded to 50 miles and an extra
section at the top of the results list is added informing the candidate
that no schools were found in their original search. This new section
has a red left border to indicate that it's not a search result.