Fixes error message anchor on schools profile show
We're using `f.check_box_input` outside of `check_box_fieldset` as we
don't want a legened for this check box. This caused us not to generate
the correct id for the containing form group div that the error message
links to.
To fix this we need to call `form_group_id` to set the id for the
containing form_group.

This PR bumps the form builder version which makes `form_group_id`
a public method.