Don't dynamically assign the imageTag
The Docker Image tag variable is dynamically assigned during the first step.

This is preventing builds from occurring - they can be manually triggered but
are not automatically firing. Best guess is the endpoint which catches the
GitHub hook is no longer able to parse the yaml and is just silently failing.
This was previously working, and since manual builds work this suggests is
just an issue with the endpoint for the GitHub hook.

Solution is to just only use a different image tag when pushing - this commit
adds a extra docker tag step and pushes the new tag instead.