Add and link to the school switch controller
The switch process is very straightforward, we remove the `current_user`
from the session and redirect the user to an internal (protected) path.
OIDC's workflow will then intercept the request and redirect to DfE
Signin - where the user is already signed in - so they will be prompted
to select an organisation.

Upon picking one and hitting 'Continue', they'll (hopefully!) be bounced
back to where we originally redirected them. Currently, this is the
dashboard.