Changed default phase to be 2
This change is only against the phase2 branch, it will land on master when we
merge phase2, at which point we will want that to be the default behaviour

DFE_SIGNIN_SECRET needed to be assigned to boot in production. That now defaults
to blank but warns the user that the environment variable is not set.