Symlink public/packs to public/packs-test
When RAILS_ENV=test webpack packs are configured to be accessed from
/public/packs-test - this doesn't exist within the container.

The Webpacker gem is hard coded to pull these values from webpack.yml and this
file cannot be dynamically changed via ERB or Env Vars.

This leaves 3 options

1. We change the file
2. We copy an alternative webpack.yml into place during the docker build process
3. We symlink the packs to packs-test during the docker build process

For Option 1 - this file will also affect developers machines so is probably a
bad idea.

For Option 2 - an alternative version of the file has the potential to drift
from the developers webpack.yml if there are any changes to the primary version.

Option 3 is a hack but simple and avoids the drift issue. It also has the
benefit that if done in ruby it will 'warm up' the bootsnap cache and be
recorded as a layer within the resultant Docker Image.