Pull school name from school not session
When we create a placement request we enqueue a job to notify the school
and the candidate and then we delete the registration_session from
redis. If the notification delivery happens quickly enough we could in
theory delete the session before the candidate has been redirected to
the show action causing them to see a session expired warning (though
this is unlikey in production).
As the only dynamic content show in the show view is the school name we
can just pull it from the url and avoid this problem entirely.