Ensure start button is at bottom of school profile
The start request button should be at the bottom of the page. Because of
our responsive layout, the right column moves beneath the left when the
page is viewed on a mobile device.

We now use CSS media queries to toggle two (identical) buttons that
ensure one is visible at the bottom of the page.