Adds school cancelling placement requests
Adds the resources to allow schools to reject a candidate's placement
request.

As per the prototype the school can preview their cancellation before
sending it. To facilitate this a cancellation now has a sent_at column
and a placement_request is only considered cancelled when it has a
cancellation and that cancellation has a non nil value for sent_at.