Don't revalidate placement request on email confirm
If a school changes it's availability preference while there are
candidates on the email confirmation step of the placement_request
wizard this will invalidate their placement_request and throw an error
when we create! the placement_request in the database. To handle this we
won't revalidate the availability before persisting the record.
If the availability was valid at the time candidate selected it then it
should still be applicable when they come back from validating their
email address. Schools still have to set a solid date when they create a
booking from the placement requests as the placement request's
availability is just a suggestion.