Disable HTML5 validations site wide
Currently when submitting a form which requires an email field to be
completed, a small non obvious error message is shown as a tooltip.

The user never sees the big obvious GDS error until they’ve successfully
passed the in browser validation. This places the accessiblity/usability
of errors under the browsers control rather than ours, and users the
particular browsers own design instead of our GDS compliant design.