Fix textarea_with_maxwords
govuk-frontend/all.js CharacterCount.prototype.createCountMessage
appends a reference to the word count span to the textarea's
aria-describedby attribute. If this attribute is not set on the textarea
the govukjs sets it to 'null' + ' ' + <id-of-word-count-span> causing an
error to be shown in waves.

Added an id to the textarea's label so we can reference the label in the
aria-describedby attribute.