Use fetch to retrieve state from session
This is an attempt to work out why the state mismatch errors are
frequently happening. Rich speculated that it might be the retrieval of
the state from Redis that's causing the issue rather than comparing the
retrieved value to the one provided in params.

Fetch will blow up if it can't find :state, potentially highlighting the
issue