Add attended flag to bookings model and scopes
Attended defaults to null giving the value three possibilities; the
attendance hasn't been recorded, the candidate has attended and the
candidate hasn't attended.

Also the previous scope isolates any bookings that should have happened
today or happened previously. When merged with the attendance_unlogged
scope we'll be able to find the records to show on the confirm
attendance page