Refactor candidate wizard state handling
Previously the knowledge about the ordering of steps in the wizard was
split between the registration session and the controller redirects.
This commit introduces a RegistrationState which determines the
next_step and a Wizard concern to handle navigation through the wizard.

NOTE: The order the candidate completes the wizard in is now important

Additionally this commit tidys up the registration session factory.