Fix back links
If we're using :back as the url and the user has clicked the 'back to
top' link at the bottom of the page they can get stuck in a loop.
Instead attempt to send the user back to where they came from or failing
that send them to the candidate start page.