Add keepalives and keepalives_interval db settings
The keepalives directive[0] controls whether or not client-side TCP
keepalives are used. The default is 1 (on). The keepalives_interval
directive[1] controls the number of seconds after which a keepalive
message should be sent.

After making these changes it's not really clear that anything has
changed, nothing special about the connection is visible in
pg_stat_activity, making thee changes tricky to test.

However, the changes can be seen by interrogating ActiveRecord's
connection info:

> ActiveRecord::Base.connection.raw_connection.conninfo

 {:keyword=>"keepalives", :val=>"1", :label=>"TCP-Keepalives", ...}
 {:keyword=>"keepalives_interval", :val=>"60", :label=>"TCP-Keepalives-Interval", ...}

Without the new settings, the keepalives and keepalives_interval
settings are present but blank - defaulting to the OS's settings

[0] https://www.postgresql.org/docs/10/libpq-connect.html#LIBPQ-KEEPALIVES
[1] https://www.postgresql.org/docs/10/libpq-connect.html#LIBPQ-KEEPALIVES-INTERVAL