Update fees edit flow to match new flow
Previously if school admins wanted to edit a fee they had to first
unselect the fee on the select fees `/schools/on_boarding/fees/new`
screen, submit the for then recheck the fee they wanted to edit.

The school on boarding wizard determines which step is next based on
whether the form model is valid, which the fees will be on update.
To have the same flow through the fees section of the wizard on both
update and new we add a boolean field for each fee to the school_profile
and check that (in addition to if the form is valid) when determining if
we need to direct the user to a screen for a fee. When updating the list
of fees offered we set the flags for the selected fees to false so the
wizard will send the user to the screens for each fee (even though they
are valid).