Add 'en-gb' culture to default Bing params
Having had a slight problem with results for the query 'Leeds' where
Bing was returning "Leeds, Kent" rather than "Leeds, Yorkshire", we were
advised by our Bing Maps support agent that:

> ... if you are querying in England it would really help if you set the
  culture to "en-gb" that will be adding "c=en-gb" to your REST Query.

We were previously using the default, which is "en-us", which was a
contributing factor to the odd search results.

Bing Maps have fixed the data issue so in that instance (for Leeds) the
correct result is returned. To reduce the risk of this problem affecting
other searches we'll also follow their advice and set the culture