Update the way of setting Chrome to be headless
This appears to have changed in a Selenium update, the 'capabilities'
approach is no longer mentioned. It has been replaced by creating an
options object and calling the `#headless!` method on it.

The other option that was originally set, disable-gpu, doesn't appear to
have a corresponding method. I suggest we allow this to remain as
default and revisit if it causes problems

https://www.rubydoc.info/gems/selenium-webdriver/Selenium/WebDriver/Chrome/Options#headless!-instance_method