Adds candidate_requirements_selection resource
The requirement is to show some % of schools the new
candidate_requirements_selection screen comrised of checkboxes and the
remaining % of schools the old candidate_requirement free text area.
To determine which requirements the school enters we set a flag on the
school profile when it is created and use this flag to determine how the
school progresses through the wizard.