Support switching devices as part of registration process
A Candidate may start the Registration process on one device (eg their laptop),
they will then expect to continue the process on the device they received their
sign in link on, potentially their phone.

We actually hold enough information to facilitate this - everything from the
first step, Personal Information, is overridden by the GiTiS Contact
information, and the school URN is included in the URN.

1. This removes some unnecessary processing of the PersonalInformation now we
instead override it with GitisRegistrationSession

2. Removes off the step check for just the token sign in since thats not
necessary in this context and it then allows the sign in to occur.