Hide DBS fee option if school doesn't require dbs
If the school chooses no for dbs requirement we set charges dbs fee to
false and don't show them the question. If they update their dbs
requirement to true we set the charges dbs field to nil to force the
user to select if they offer dbs.

This logic should eventually be moved into current step.