Add safeguards around current_user_full_name method
The current_user_full_name is ordinarily only used after we've checked
that the user is present (in the Phase Banner partial). The one other
place it's used is on the not registered error page. That call has been
wrapped to prevent the paragraph from being shown if the user is absent
and the method will now return 'Unknown' if the current user isn't a
valid UserInfo object