Only check school has fixed dates on create
When the placement request is created on a school that has flexible
dates and the school subsequently changes to fixed, the saving context
around the presence of a bookings placement date changes rendering the
existing placement requests invalid.

This problem hasn't reared its head but now we are updating the
viewed_at timestamp it's possible to trigger it, as the placement
request object is invalid.

Ensuring that either the availability description or placement date
being present should be sufficient and both directions are accounted for