Stop describe_current_search capitalising searches
Since the initial implementation of search the way the search is
described has improved to the point where the original location isn't as
important as it was (now we show the result from Geocoder rather than
the search term in most cases).

To improve the experience when searching for postcodes, the initial
query is no longer capitalised when a match isn't found. This should be
less confusing for the person searching as it will match exactly what
they typed in.

Additionally, now we no longer support search near my location, that
functionality has been removed from the description method and
associated specs