Add supports_subjects to phases, placement dates
This flag is intended to act as an identifer on whether or not to
present school users with the option to associate subjects with
placement dates.

For phases that don't "support subjects", like Primary, no
subject-specificity should be displayed.

This option will be presented to the user when creating placement dates
that categorise the date as having subjects (primary) or not (secondary,
college) when schools offer multiple phases