First attempt at actually presenting the selector
This is slightly complicated because we are trying to select two things
with a single radio button.

The collection is divided into subgroups by day and each subject
available on that day is listed as an actual radio button, but because
subjects can appear more than once (even in a subgroup) we need to
submit both IDs.

For the time being, they are formatted in the following manner

=> 1_2  -   where 1 is the :bookings_placement_date_id,
	    and 2 is the :bookings_placement_date_subject_id

In the case of 'All subjects', no subject is appended and only the
:bookings_placement_date_id is set as the value

These will be parsed and separated out by the controller.

This is clearly a WIP and needs more attention