Removes nesting from registration session
Stores registrations in a flat structure in the session rather than
nesting attributes under a model name.
Delegates how to instantiate a step and what attributes to store to the
RegistrationStep.
All registration_step attribute names, except for time stamps, are
unique so we don't need to name space them to avoid collisions. Only
time stamps will collide so we need to  prefix those with the models
name.