Handle in flight legacy sessions.
In flight registrations for fixed date schools need to handle the new
subject specific code.
If the candidate has completed the placement_preference step for a fixed
date school the placement_preference attributes will contain a
bookings_placement_id, we can use this to identify legacy sessions.
We can treat a legacy session for fixed date school as the candidate
having applied for a non subject_specific date and build a valid
subject_and_date_information_step. This will allow the candidate to
continue and also allow any background jobs to work.

bookings_placement_date_id has been readded to placement_preference to
preserve backwards compatability. It has been added as an attr_accessor
so it doesn't show up in the models attributes and overwrite the
bookings_placement_date_id from subject_and_date_information_step when
converting the session to a placement request in
registration_session_as_placement_request.

After 24 hours the changes in this commit can be removed.