Adds new_registration_session_factory
Adds a new registration_session_factory that stores data in the same
structure as the refactored registration_session.
Once the registration_session refactoring has been live for 24 hours
all the specs can be updated to use the new_registration_session_factory
and the old registration_session_factory can be removed.