Log accordion toggling to Google Analytics
This approach monkey patches the accordion's setExpanded function
(thanks @jebw!) and, if there's a data-track-event-name attribute,
records the event to Google Analytics.

Currently the only accordion logged is the DfE Sign-in help on the
school's home page, adding data-track-event-name to other accordions
should work

Note, if accordion segments are left open, revisiting the page will
count as a view even if the user hasn't interacted with the accordion on
that visit.