Feature flag has_limited_availability
The do you have limited availability section of the placement_date form
was added as it was in the prototype at the time of implementing subject
specific dates, however it's a seperate piece of work.  This has been
feature flagged off, until the canidate side of the work has been done.
All placement dates will have no limit to the maximum number of
bookings, until the feature is enabled.