Fixes school changing date placement request bug
If a placement_request is made for a flexible date and the school changes
their dates to fixed the placement request will be invalid. To avoid
this, for the time being, make the validations contextual.
The issue arises from placement_request and registration_session sharing
this validation, it should be split out.