Don't invalid placements_request on date changes
If a school changes a date to be subject_specific then we'll invalid any
existing placement_requests for that date as they'll now be expected to
have a bookings_placement_dates_subject_id.
Make the validation contextual as a workaround. The issue arises as the
validation is shared with a registration step, the validations should be
seperated as a better fix.