Add subject to placement_requests
If a placement request were made for a subject specific date and the
school attempted to remove the subject from that date the app would
throw a referential integrity error in postgres. Removing the subject
from the date deletes the bookings_placement_dates_subject our
placement_request has a foreign_key for.

To fix this we no longer store the bookings_placement_dates_subject on
the placement_request. It's now removed in
RegistrationAsPlacementRequest. We now store the bookings_subject_id
from the subject specific date (set in the
subject_and_date_information_step).