Standardise subject and date info controller
Updates the subject and date information to match the other
registrations controllers.  We can have `subject_and_date_ids=` do all
the lifting of splitting the param and finding the right date and
subject (if applicable) then we can just pass the params straight in
like a normal rails controller.