Stop raising errors for session expiry
Sometimes, for reasons unknown, users reach our OIDC callback path
`/auth/callback` with the param `?error=sessionexpired`. We previously
treated this as an AuthFailedError but that was causing too much noise
in our error reporting.

Having asked on the DfE Sign-in Slack channel and getting no response I
had a search on GitHub and other teams are manually handling this
error[0][1].

This PR implements that approach, users who are redirected back with the
sessionexpired error, instead of raising AuthFailedError will raise
SessionExpiredError. This is handled by session_expired_failure which
redirects instead of reporting the error to Sentry via Raven.

Users will be redirected to `/schools/dashboard` which will, in turn,
send them back to DfE Sign-in where they will log in again.

[0] https://github.com/DFE-Digital/manage-courses-frontend/blob/master/config/initializers/omniauth.rb#L19
[1] https://github.com/DFE-Digital/login.dfe.support/blob/master/src/infrastructure/oidc/index.js#L74