Fix publishing profile for schools w/o dbs
We removed the requirement for schools to provide an explanation when
they don't require dbs checks on the on boarding wizard. However we
didn't update the validation on bookings profile to account for this.
This allowed schools to complete the onboarding wizard but error when
attempting to publish their profiles.