Disable forgery protection for schools namespace
We've been getting a number of Invalid authenticity token errors on
production due to a null origin header being present.
Looking in sentry this seems to have only effected schools, the most
likely explanation is due to their firewall settings or some type of do
not track plugin.
Disabling this check so schools can continue to use the service.