Fixes device switching for subject specific dates
After we persist a step to the registration session write the updated
registration_session to redis. On each request we attempt to find the
existing registration_session in redis and point the user's session at
that.
The user's browser session now contains a pointer to the
registration_session stored in redis rather than containing the
registration.

```
browser_session => {
  'registrations' => {
    'urn' => 'registration_session_uuid'
  }
}

redis => {
  'registration_session_uuid' => { registration_session_data }
}
```

When the user comes back from email verification we pull the pointer to
the correct registration_session from the url params and update their
browser session to include this pointer, then when current_registration
is next called their old session is loaded from redis, saving them from
having to redo already completed steps.