Use development branch images for docker cache
Currently if dependencies get updated on master, a new cached base image is
created but if the are updated on development, than one is not.

Normally dependencies are updated on master but if a new version of the form
builder is merged to development than development, and subsequent branches, will
remain uncached until the dependency change makes it out to master.