Ensure the params match the bookings
If the user hits the back button in the browser then resubmits the form
then some of the bookings in the params might not be in the
unlogged_bookings collection, causing an error to be thrown in the
School::Attendance object.

Reject any params that don't correspond to an unlogged_booking.