Add reminder schedules, tests and plumb in sending
The one week reminders are scheduled to be enqueued at 02:35 and the one
day ones at 02:30. Doing a single query for bookings and queueing an
email job for each of the _should_ run within the 3 minute timeframe that
DelayedJob offers.

The queueing part is wrapped in a transaction so unless all reminders are
queued within the limit none will be. If our load increases to a point
where all emails can't be queued within 3 minutes (several thousand
bookings on a single day) this approach will have to be changed so they
are done in batches.