Patch ActiveRecord::RecordInvalid exception to bypass translations
Currently we are loosing the fieldnames in the RecordInvalid exceptions which
makes it hard to debug the source of the error.

The exception messages is not surfaced to the user so changed the definition of
RecordInvalid to include a lower level representation of the error messages
with more complete information.