Replace generic back link to link to date index
If an invalid submission is made this change allows users to still go
'back' to the date index. This does create the possibility of 'losing'
placement dates that have been created but not configured, but as the
date has not yet been _confirmed_ to the user they shouldn't expect it
to 'persist' (which it does, it's just not surfaced anywhere).