Fix for https://github.com/sass/sassc-ruby/issues/146
The sassc gem compiles using -march=native which produces a non-portable c lib
of sassc. In a Docker environment, the built image may be run on a different
system than the one it is compiled on - potentially resulting in a segfault.

This looks to be the issue we have been encountering over the last week.

2.2.1 is supposed to fix this but does not actually remove the -march=native
optimisation and I believe continues to exhibit this problem.

NOTE: the default on `master` in the sassc-ruby library is now false so whenever
the next release is made, this change can be dropped