Changed how 'Accept booking' works
Previously we had two controllers, ConfirmBooking and MakeChanges both
processing Bookings for acceptance. Now clicking Continue on ConfirmBooking with
prefilled values creates a hidden for and submits to MakeChangesController.

1. This avoids two controllers processing the same data
2. Previously ConfirmBooking didn't handle failed validations properly and would
just reshow the page with a 'Continue' button to hop across to MakeChanges. Now
it will instead POST to MakeChanges, fail to save and show an appropriate error
message to the user.