Simplified Dfe Sign-in role check
Previously this was doing its own translation (and caching) of the Schools
organisation id. Instead use the organisation id provided by DfE Sign-in and
avoid caching these in the database.

This resolves a bug where if the DfE Sign-in user belonged to organisations
which were not part of the service, than the database could not update the
cached copy of the organisations uuid.