Skip verify_authenticity_token for cookie preferences
If the user has left the page open for an extended period of time, then it is
likely the token will expire. If they then return and click 'accept' they will
see an unexpected session expired message.

The update action only sets a cookie on the users web browser so it is safe to
not check the authenticity token, I am disabling it to avoid the user seeing
the error unnecessarily.