Changed redirection path when the user is not associated with an organisation
Prior behaviour cast the blank school urn #to_i, which resulting in a 0. This
meant the user saw the 'School not registered' error instead of the No school
error which appears to be a stub screen.

Instead redirect the user to the insufficient privileges screen which has more
useful information on it for resolving the users issues.