Removed support for invalid config
The recent change to use RoleCheckedOrganisations means we always use role
checks if the API is enabled. A subsequent change to make the Roles#enabled?
return false if the config is disabled breaks the API usage so removing the
option to toggle just this specific API.

The actual issue is Roles#enabled? will now trigger a ApiDisabled error in the
attempt to call response in parent DFESignInAPI::Client class.