Add Flipper to manage feature switches
We are migrating the app to use the GiT API in order to interface
with the Dynamics CRM. Currently, we go to the Dynamics CRM directly
- migrating is going to be reasonably complex and multi-stage
  process, so it would be sensible to keep the current implementation
as a fallback should anything unexpectadly go wrong during the
migration process.

Feature switches will enable us to quickly fallback to the original
CRM integration in the event of an issue being discovered in
production.