Run Flipper against memory store in test env
We use the same Redis instance for our unit tests as development, which causes
the features to be enabled in test that are enabled in dev.

Instead, we want to by default have all features off in test - we achieve this
by setting Flipper to use an in-memory cache store for the test environment.