Update registration mapper to support API model
When we start pulling candidates back from the API instead of
Dynamics directly they will be an instance of
`SchoolsExperienceSignUp` and have slight different attribute
names to the `Bookings::Gitis::Contact` that we are replacing.

Some of the domain logic in the existing
`Bookings::Gitis::Contact` has been moved to the mapper model
for now; this is to avoid adding _another_ intermediary model
that abstracts the `SchoolsExperienceSignUp`. Going forward I
think I can push much of this behaiour into the API.

Defaults the accepted privacy policy for the mapped contact -
currently this is a separate operation at the controller-level,
but with the API the model encapsulates the accepted policy as
well.