Update Candidate model to support API integration
The `Bookings::Candidate` model contains various methods to find
and update candidate records based on the associated
`gitis_contact`. Until now, the `gitis_contact` has always been
`Bookings::Gitis::Candidate`, however, going forward the API
will be specifying a `contact` that is an instance of
`GetIntoTeachingApiClient::SchoolsExperienceSignUp`.

For the most part we just need to ensure we call `candidate_id`
if the object doesn't respond to `id` (the API model uses the
`candidate_id` attribute name).

This `Bookings::Candidate` class is also responsible for mapping
a registration to a contact instance and persisting it to the
CRM. In this case, we check for the `git_api` feature flag and
call the API when enabled. It is fine if the state is such that
`contact` is an instance of `Bookings::Gitis::Candidate` and we
are now sending it to the API as the support is
backwards-compatible with this model type as well (the mapper
doesn't work of the feature flag so should adapt to the input
given).