Support issuing verification codes
When the personal information of a candidate is constructed in
the controller we currently create a session and generate a
token that is emailed to the user.

Instead, we need to be able to call the API with the personal
information of the user in order to issue a verification code.
If the API throws any kind of error we treat it as not possible
to matchback the candidate and return `false`. If the candidate
was successfully matched back and emailed with a verification
code we return `true`; this will enable us to serve the correct
subsequent view in the controller.