Issue verification code via API
When the `git_api` feature is enabled we want to bypass the
current session creation and associated magic link
generation/delivery to instead call the API to generate and
email a verification code.

The flow remains the same, in that if the candidate is
successfully matched back we will still take them to the sign in
path (we will need to update this to support entering the
verification code).